% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cases.R
\name{get_cases}
\alias{get_cases}
\title{COVID-19 cases scraper}
\usage{
get_cases(news)
}
\arguments{
\item{news}{news links data frame, output of \code{get_news()} or a subset of that data frame}
}
\value{
A data frame with self-explanatory columns.
}
\description{
Scrapes data on the number of new COVID-19 cases conducted in Russian regions from Rospotrebnadzor website
}
\examples{
# get all the data
df <- get_cases(get_news())
}
