% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tests_federal.R
\name{get_tests_federal}
\alias{get_tests_federal}
\title{Federal tests scraper}
\usage{
get_tests_federal(news)
}
\arguments{
\item{news}{news links data frame, output of \code{get_news()} or a subset of that data frame}
}
\value{
A data frame with self-explanatory columns.
}
\description{
Scrapes data on the cumulative number of COVID-19 tests conducted in Russia from Rospotrebnadzor website
}
\examples{
# get all the data from the first page of news
news <- get_news(pages = 1)
df <- get_tests_federal(news)
}
