% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen.R
\docType{data}
\name{old_cases}
\alias{old_cases}
\title{Old data on cases}
\format{
A data frame
\describe{
\item{region}{name of the region}
\item{cases}{number of cases reported that day}
\item{date}{date}
}
}
\source{
\url{https://www.rospotrebnadzor.ru/}
}
\usage{
old_cases
}
\description{
A data frame with the data on cases from March 5th to March 18th plus March 23rd that weren't reported as tables
}
\keyword{datasets}
