% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_news.R
\name{get_news}
\alias{get_news}
\title{News scraper}
\usage{
get_news()
}
\value{
A data frame with self-explanatory columns.
}
\description{
Scrapes links to news from the Rospotrebnadzor website
}
\examples{
# get the news
df <- get_news()
}
